Add Kafka Consumer to Notification Service
This commit introduces the following changes to the codebase:

- Create `NotificationService` class
- Add `@KafkaListener` annotation to `listen` method in `NotificationService`
- Configure the `listen` method to consume messages from the `notificationTopic` Kafka topic with the group ID `notification-service-group`
- Log the received order ID from the Kafka message

Purpose:
The purpose of this commit is to implement a Kafka consumer in the Notification Service application. The `NotificationService` class contains a method annotated with `@KafkaListener` that listens for messages on the `notificationTopic` Kafka topic. When a message is received, the `listen` method is invoked, and it logs the order ID received from the message. This allows the Notification Service to receive and process order notifications published by the Order Service to the `notificationTopic` Kafka topic.